extract `mem_replace_option_with_none` tests into their own file
Note that this doesn't migrate the no-std test for issue9824. The
motivation is that the no-std file is usually supposed to just test that
the basic functionality of the lint works in no-std, which boils down to
`std` being properly replaced with `core`. Bringing a reproducer for a
separate issue thus doesn't add much value.
