+++ tests/ui/mem_replace.fixed
@@ -1,20 +1,7 @@
-#![warn(
-    clippy::mem_replace_option_with_none,
-    clippy::mem_replace_option_with_some,
-    clippy::mem_replace_with_default
-)]
+#![warn(clippy::mem_replace_option_with_some, clippy::mem_replace_with_default)]
 
 use std::collections::{BTreeMap, BTreeSet, BinaryHeap, HashMap, HashSet, LinkedList, VecDeque};
 use std::mem;
-
-fn replace_option_with_none() {
-    let mut an_option = Some(1);
-    let _ = an_option.take();
-    //~^ mem_replace_option_with_none
-    let an_option = &mut Some(1);
-    let _ = an_option.take();
-    //~^ mem_replace_option_with_none
-}
 
 fn replace_with_default() {
     let mut s = String::from("foo");
@@ -116,37 +103,13 @@ fn msrv_1_40() {
 }
 
 fn issue9824() {
-    struct Foo<'a>(Option<&'a str>);
-    impl<'a> std::ops::Deref for Foo<'a> {
-        type Target = Option<&'a str>;
-
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl<'a> std::ops::DerefMut for Foo<'a> {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
-        }
-    }
-
     struct Bar {
-        opt: Option<u8>,
         val: String,
     }
 
-    let mut f = Foo(Some("foo"));
     let mut b = Bar {
-        opt: Some(1),
         val: String::from("bar"),
     };
-
-    let _ = f.0.take();
-    //~^ mem_replace_option_with_none
-    let _ = (*f).take();
-    //~^ mem_replace_option_with_none
-    let _ = b.opt.take();
-    //~^ mem_replace_option_with_none
 
     let _ = std::mem::take(&mut b.val);
     //~^ mem_replace_with_default
