+++ tests/ui/mem_replace_no_std.rs
@@ -1,20 +1,7 @@
-#![warn(
-    clippy::mem_replace_option_with_none,
-    clippy::mem_replace_option_with_some,
-    clippy::mem_replace_with_default
-)]
+#![warn(clippy::mem_replace_option_with_some, clippy::mem_replace_with_default)]
 #![no_std]
 
 use core::mem;
-
-fn replace_option_with_none() {
-    let mut an_option = Some(1);
-    let _ = mem::replace(&mut an_option, None);
-    //~^ mem_replace_option_with_none
-    let an_option = &mut Some(1);
-    let _ = mem::replace(an_option, None);
-    //~^ mem_replace_option_with_none
-}
 
 fn replace_with_default() {
     let mut refstr = "hello";
@@ -37,35 +24,11 @@ fn dont_lint_primitive() {
 }
 
 fn issue9824() {
-    struct Foo<'a>(Option<&'a str>);
-    impl<'a> core::ops::Deref for Foo<'a> {
-        type Target = Option<&'a str>;
-
-        fn deref(&self) -> &Self::Target {
-            &self.0
-        }
-    }
-    impl<'a> core::ops::DerefMut for Foo<'a> {
-        fn deref_mut(&mut self) -> &mut Self::Target {
-            &mut self.0
-        }
-    }
-
     struct Bar {
-        opt: Option<u8>,
         val: u8,
     }
 
-    let mut f = Foo(Some("foo"));
-    let mut b = Bar { opt: Some(1), val: 12 };
-
-    // replace option with none
-    let _ = mem::replace(&mut f.0, None);
-    //~^ mem_replace_option_with_none
-    let _ = mem::replace(&mut *f, None);
-    //~^ mem_replace_option_with_none
-    let _ = mem::replace(&mut b.opt, None);
-    //~^ mem_replace_option_with_none
+    let mut b = Bar { val: 12 };
     // replace with default
     let _ = mem::replace(&mut b.val, u8::default());
 }
