+++ tests/ui/mem_replace_no_std.stderr
@@ -1,20 +1,5 @@
-error: replacing an `Option` with `None`
-  --> tests/ui/mem_replace_no_std.rs:12:13
-   |
-LL |     let _ = mem::replace(&mut an_option, None);
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider using `Option::take()` instead: `an_option.take()`
-   |
-   = note: `-D clippy::mem-replace-option-with-none` implied by `-D warnings`
-   = help: to override `-D warnings` add `#[allow(clippy::mem_replace_option_with_none)]`
-
-error: replacing an `Option` with `None`
-  --> tests/ui/mem_replace_no_std.rs:15:13
-   |
-LL |     let _ = mem::replace(an_option, None);
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider using `Option::take()` instead: `an_option.take()`
-
 error: replacing a value of type `T` with `T::default()`
-  --> tests/ui/mem_replace_no_std.rs:21:13
+  --> tests/ui/mem_replace_no_std.rs:8:13
    |
 LL |     let _ = mem::replace(&mut refstr, "");
    |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider using `core::mem::take` instead: `core::mem::take(&mut refstr)`
@@ -23,28 +8,10 @@ LL |     let _ = mem::replace(&mut refstr, "");
    = help: to override `-D warnings` add `#[allow(clippy::mem_replace_with_default)]`
 
 error: replacing a value of type `T` with `T::default()`
-  --> tests/ui/mem_replace_no_std.rs:25:13
+  --> tests/ui/mem_replace_no_std.rs:12:13
    |
 LL |     let _ = mem::replace(&mut slice, &[]);
    |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider using `core::mem::take` instead: `core::mem::take(&mut slice)`
 
-error: replacing an `Option` with `None`
-  --> tests/ui/mem_replace_no_std.rs:63:13
-   |
-LL |     let _ = mem::replace(&mut f.0, None);
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider using `Option::take()` instead: `f.0.take()`
-
-error: replacing an `Option` with `None`
-  --> tests/ui/mem_replace_no_std.rs:65:13
-   |
-LL |     let _ = mem::replace(&mut *f, None);
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider using `Option::take()` instead: `(*f).take()`
-
-error: replacing an `Option` with `None`
-  --> tests/ui/mem_replace_no_std.rs:67:13
-   |
-LL |     let _ = mem::replace(&mut b.opt, None);
-   |             ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^ help: consider using `Option::take()` instead: `b.opt.take()`
-
-error: aborting due to 7 previous errors
+error: aborting due to 2 previous errors
